+++ realm/realm-library/src/objectServer/java/io/realm/internal/network/AuthServerResponse.java
@@ -63,7 +63,14 @@ public abstract class AuthServerResponse {
             JSONObject obj = new JSONObject(response);
             String title = obj.optString("title", null);
             String hint = obj.optString("hint", null);
-            ErrorCode errorCode = ErrorCode.fromInt(obj.optInt("code", -1));
+            ErrorCode errorCode;
+            if (obj.has("code")) {
+                errorCode = ErrorCode.fromInt(obj.getInt("code"));
+            } else if (obj.has("status")) {
+                errorCode = ErrorCode.fromInt(obj.getInt("status"));
+            } else {
+                errorCode = ErrorCode.UNKNOWN;
+            }
             return new ObjectServerError(errorCode, title, hint);
         } catch (JSONException e) {
             return new ObjectServerError(ErrorCode.JSON_EXCEPTION, "Server failed with " +